requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.26.1=pypi_0
anyio=3.6.2=pypi_0
appdirs=1.4.4=pypi_0
appnope=0.1.3=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.1.0=pypi_0
attrs=22.1.0=pypi_0
babel=2.11.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=openblas
bleach=5.0.1=pypi_0
brokenaxes=0.6.1=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.5.7=hbcca054_0
cachetools=5.3.1=pypi_0
cairo=1.16.0=hb05425b_4
certifi=2022.9.24=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
comm=0.1.0=pypi_0
contourpy=1.0.6=pypi_0
cycler=0.11.0=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
diffusers=0.27.2=pypi_0
docker-pycreds=0.4.0=pypi_0
eigen=3.4.0=h4bd325d_0
entrypoints=0.4=pypi_0
executing=1.2.0=pypi_0
expat=2.2.10=h9c3ff4c_0
fastjsonschema=2.16.2=pypi_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.13.1=pypi_0
flatbuffers=1.12=pypi_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.38.0=pypi_0
freetype=2.10.4=h0708190_1
fsspec=2023.12.2=pypi_0
ftfy=6.1.3=pypi_0
giflib=5.2.1=h36c2ea0_2
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
glib=2.69.1=he621ea3_2
glide-text2im=0.0.0=dev_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
grad-cam=1.4.8=pypi_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.7.0=pypi_0
harfbuzz=4.3.0=hf52aaf7_1
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.20.2=pypi_0
icu=58.2=hf484d3e_1000
idna=3.4=pypi_0
imageio=2.22.4=pypi_0
importlib-metadata=7.0.1=pypi_0
ipykernel=6.18.0=pypi_0
ipython=8.6.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.2=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h166bdaf_1
json5=0.9.10=pypi_0
jsonschema=4.17.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.4.7=pypi_0
jupyter-console=6.4.4=pypi_0
jupyter-core=5.0.0=pypi_0
jupyter-server=1.23.3=pypi_0
jupyterlab=3.5.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.16.3=pypi_0
jupyterlab-widgets=3.0.3=pypi_0
kaleido=0.2.1=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=pypi_0
kornia=0.6.8=pypi_0
krb5=1.19.3=h3790be6_0
lab=7.2=pypi_0
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=15_linux64_openblas
libcblas=3.9.0=15_linux64_openblas
libclang=14.0.6=pypi_0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.12=h8f2d780_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libgomp=11.2.0=h1234567_1
liblapack=3.9.0=15_linux64_openblas
libllvm10=10.0.1=he513fc3_3
libopenblas=0.3.20=pthreads_h78a6416_0
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=h6a678d5_2
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
llvmlite=0.40.1=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
mpmath=1.3.0=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.1=pypi_0
nbconvert=7.4.0=pypi_0
nbformat=5.9.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
nettle=3.6=he412f7d_0
networkx=3.1=pypi_0
notebook=6.5.4=pypi_0
notebook-shim=0.2.3=pypi_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numba=0.57.1=pypi_0
numpy=1.23.5=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-ml-py=12.535.108=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.99=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
nvitop=1.3.0=pypi_0
oauthlib=3.2.2=pypi_0
opencv=4.6.0=py310h1128e8f_3
opencv-python=4.8.0.76=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1t=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pandas=1.5.1=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.3.0=pypi_0
pip=23.0.1=py310h06a4308_0
pixman=0.40.0=h36c2ea0_0
platformdirs=2.5.4=pypi_0
plotly=5.14.1=pypi_0
prometheus-client=0.15.0=pypi_0
prompt-toolkit=3.0.33=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pygments=2.13.0=pypi_0
pynndescent=0.5.10=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.2=pypi_0
python=3.10.11=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python_abi=3.10=2_cp310
pytz=2022.6=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=24.0.1=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtconsole=5.4.0=pypi_0
qtpy=2.3.0=pypi_0
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.28.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.1=pypi_0
scienceplots=2.0.1=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.3=pypi_0
scipy=1.9.3=pypi_0
send2trash=1.8.0=pypi_0
sentry-sdk=1.22.2=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=67.8.0=py310h06a4308_0
simplejson=3.18.0=pypi_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.1=pypi_0
sympy=1.12=pypi_0
sync-batchnorm=0.0.1=pypi_0
tenacity=8.2.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-estimator=2.9.0=pypi_0
termcolor=2.1.1=pypi_0
terminado=0.17.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.10.10=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.0=pypi_0
tomli=2.0.1=pypi_0
torch=2.2.1=pypi_0
torchvision=0.17.1=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.5.0=pypi_0
transformers=4.36.2=pypi_0
triton=2.2.0=pypi_0
ttach=0.0.3=pypi_0
txt2tags=3.7=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2023c=h04d1e81_0
umap-learn=0.5.3=pypi_0
urllib3=1.26.12=pypi_0
wandb=0.15.2=pypi_0
wcwidth=0.2.12=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.4.2=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.38.4=py310h06a4308_0
widgetsnbextension=4.0.3=pypi_0
wrapt=1.14.1=pypi_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.2=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
